requirements.txt
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
email-validator
python-multipart
jinja2